update searcher and openvino launcher
