+++ datumaro/plugins/openvino_plugin/samples/clip_text_ViT-B_32_interp.py
